+++ src/main/java/roomescape/controller/web/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.file;
+package roomescape.controller.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
